changed the naming and project path
