--- primary_relationships/src/main/resources/relationships.conf
@@ -1,17 +1,0 @@
-cassandra {
-  hosts = "127.0.0.1"
-  username = "user"
-  password = "pwd"
-  keyspace = "keyspace"
-  knowledge_base_table = "knowledge_base_table"
-  mv_feature_type_table = "mv_feature_type_table"
-  primary_relationships_table = "primary_relationships_table"
-  results = "results"
-  results_meta_by_sha256 = "results_meta_by_sha256"
-  results_data_by_sha256 = "results_data_by_sha256"
-}
-
-spark {
-  appName = "relationships"
-  master = "localhost"
-}
