+++ primary_relationships/src/main/scala/com/holmesprocessing/analytics/relationship/models/Knowledge_Base.scala
@@ -1,4 +1,4 @@
-package com.holmesprocessing.relationships.models
+package com.holmesprocessing.analytics.relationship.models
 
 import java.util.UUID
